# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EpiCare AI",
    "description": "\n**EpiCare AI** is a robust and comprehensive platform designed to manage and analyze epilepsy-related data, facilitating seamless interactions between patients, doctors, and caretakers.\n\n### Core Features\n- **Admin Dashboard**: Comprehensive system metrics and doctor PMDC verification tools.\n- **User Management**: Unified profiles tailored for Patients, Doctors, and Caretakers.\n- **Role-Based Authentication**: Secure access control with PMDC verification for doctors.\n- **Connection System**: Connect patients with verified medical professionals and trusted caretakers.\n- **AI-Driven EEG Analysis**: Upload EEG sessions, process signals, and receive detailed AI-generated seizure prediction reports.\n- **Personalized Dashboard**: Get real-time heuristic recommendations based on daily habits.\n- **Lifestyle & Medication Tracking**: Log sleep, stress, triggers, and medications with adherence algorithms.\n- **Emergency SOS**: Instantly dispatch multi-channel (Firebase/WhatsApp/Email) SOS alerts to configured contacts.\n\nThis API adheres strictly to REST principles, delivering standardized responses and explicit HTTP status codes to ensure a reliable developer experience.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List all users",
        "description": "Retrieve all users with optional pagination and role filtering.",
        "operationId": "list_users_api_v1_admin_users_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/UserRole"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserOut"
                  },
                  "title": "Response List Users Api V1 Admin Users Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{user_id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get user details",
        "description": "Fetch a specific user's base information by their ID.",
        "operationId": "get_user_api_v1_admin_users__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{user_id}/status": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "summary": "Activate or Deactivate User",
        "description": "Suspend or restore a user account.",
        "operationId": "update_user_status_api_v1_admin_users__user_id__status_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/doctors/pending": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List pending doctors",
        "description": "Fetch all doctors whose PMDC numbers are awaiting verification.",
        "operationId": "get_pending_doctors_api_v1_admin_doctors_pending_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DoctorProfileOut"
                  },
                  "type": "array",
                  "title": "Response Get Pending Doctors Api V1 Admin Doctors Pending Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/doctors/{user_id}/verify": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "summary": "Verify Doctor PMDC",
        "description": "Approve or reject a doctor's PMDC verification request.",
        "operationId": "verify_doctor_api_v1_admin_doctors__user_id__verify_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DoctorVerificationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DoctorProfileOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "pmdc_number": {
            "type": "string",
            "title": "Pmdc Number"
          },
          "specialty": {
            "type": "string",
            "title": "Specialty"
          },
          "hospital_affiliation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hospital Affiliation"
          },
          "license_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Image Url"
          },
          "is_pmdc_verified": {
            "type": "boolean",
            "title": "Is Pmdc Verified"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "pmdc_number",
          "specialty",
          "is_pmdc_verified",
          "created_at",
          "updated_at"
        ],
        "title": "DoctorProfileOut"
      },
...
      "DoctorVerificationUpdate": {
        "properties": {
          "is_verified": {
            "type": "boolean",
            "title": "Is Verified",
            "description": "Set to true to verify the doctor, false to reject."
          }
        },
        "type": "object",
        "required": [
          "is_verified"
        ],
        "title": "DoctorVerificationUpdate",
        "description": "Schema to verify or reject a doctor's PMDC number."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "phone_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone Number"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "is_email_verified": {
            "type": "boolean",
            "title": "Is Email Verified"
          },
          "is_phone_verified": {
            "type": "boolean",
            "title": "Is Phone Verified"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "phone_number",
          "full_name",
          "role",
          "is_active",
          "is_email_verified",
          "is_phone_verified",
          "created_at",
          "updated_at"
        ],
        "title": "UserOut",
        "description": "Response model (never exposes password)."
      },
...
      "UserRole": {
        "type": "string",
        "enum": [
          "PATIENT",
          "DOCTOR",
          "CARETAKER",
          "ADMIN"
        ],
        "title": "UserRole"
      },
      "UserStatusUpdate": {
        "properties": {
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Set to false to suspend the user account."
          }
        },
        "type": "object",
        "required": [
          "is_active"
        ],
        "title": "UserStatusUpdate",
        "description": "Schema to activate or deactivate a user account."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}